requirements.txt
dearpygui==1.11.1
keras==3.4.1
numpy==1.26.4
opencv-python==4.10.0.84
tensorflow==2.17.0
pillow==10.4.0
